Execute routes off a stream instead of a projection.
The semantics of calling a function with side-effects on a projection
are undefined.  The projection worked in Scala 2.7, but in Scala 2.8,
the actions are called three times each.

A stream is safe because a stream's contents are memoized.  This means
that actions only execute once.

Making the change on the master branch, just to minimize the differences
between master and scala-2.8.