Automate the synchronization between sbt build and POMs.
No changes should be made directly to the POMs.  Any changes should be
made to ScalatraProject, and the POMs regenerated with 'sbt make-pom'.
This should give us the best of both tools with a single point of
maintenance.

Known issue: the Maven build does not run any tests.  Need to either
integrate unreleased maven-scalatest-plugin or sprinkle @JUnitRunner
annotations everywhere.