Publish the parent POM from SBT.
For reasons still unclear, the generated POMs had to be removed or else
SBT would silently fail to publish the parent POM.  Debug showed it
attempting, in vain, to load a JAR for the parent project, but it wasn't
clear why.  Whatever.  It works now, and the generated POMs were going
away anyway.