[breaking] Make trailing slash optional on non-root path-mapped servlets. [fixes #98]
This only breaks if you are using a ScalatraServlet and its mapping does
not end in `/*`.

We used to identify non-path-mapped servlets by a null pathInfo.  A
path-mapped servlet, other than `/*`, also returns a null pathInfo when
the servlet mapping is requested without a trailing space.  The latter
case is a lot more common than the former, so we are going to optimize
for that.

If your routes suddenly stopped working, try:

    override def requestPath = request.getServletPath