ScalatraFilter's methodNotAllowed call filter chain; also allow pass.
If a filter matches a route for a different method, but a nested filter
can handle the request, it wasn't getting called.  This lets it fall
through to the filter chain.  Note that we can't reliably tell whether
the filter chain can handle the request, so this makes 405 detection
impossible in a ScalatraFilter.  ScalatraServlet is unaffected.