Bootstrap Scalatra from a LifeCycle trait. [fixes #81]
Add the following to web.xml.  XMLophobes may use the JettyServer.

    <listener>
      <listener-class>org.scalatra.servlet.ScalatraListener</listener-class>
    </listener>

The server will then look for a class named Scalatra in the default package.
This class should extend the LifeCycle trait.  That trait can be used to
mount servlets and filters and do whatever other bootstrapping is necessary
in your application.