Infer content type from the headers map given with ActionResult
Previously if Content-Type header was set in the headers map for ActionResult,
it would be discarded and overriden by the header inferred using
contentTypeInferrer.

This change tries to infer the Content-Type from the headers map for
ActionResults, or if it isn't specified, falls back to inferring the
content type from the body.