Make AuthOperation compile.  Fixes GH-276
This was introduced in b6934902a3aec3352b9fe57a78f4df21ecd59ff8.  The
fix here could break uses of the AuthOperation case class.  The risk
seems slight, but we try hard not to break API in patch releases.
Needs review by a Swaggerer, but this unbreaks the build.