use body method to read the body of the request
With the following commit, it was changed to read the cached contents with
the body method rather than reading the body part independently from inputStream.

https://github.com/scalatra/scalatra/commit/a9f092ca69295eb87a9bd7247ca3730bc552c46c

Once inputStream is consumed once, it can not be read again, so the acquisition
of the body part will be unified with the body method of the request object.

The body method of the request object considers the body of JSON type,
and guarantees that it is called multiple times.