Fixed not to get JSON data from parameters
When the Content-Type is 'application/x-www-form-urlencoded',
the data of the body part is set as the parameter of the Sevlet
specification.

When Content-Type is 'application/x-www-form-urlencoded', there
is prepared a code which obtains JSON data from parameters, but
in the first place there is a code whose Content-Type is "json"
This conditional branch will never be true as it will not be
reached at the time of "xml".