deterred the premature JSON parse
In the 'invoke' method, it is judged whether or not parsing is
necessary with the 'shouldParseBody' method, but since you are
calling 'parsedBody' in the 'shouldParseBody' method in the
first place, if you mix-in 'JsonSupport', every time an action
is executed, JSON's Parsing necessity judgment and execution of
parsing itself are done.

This should be executed explicitly in the action.