changed the way to guess the MIME Type of a file
Changed to use URLConnection.guessContentTypeFromName
for guessing the MIME Type from the file path.

By doing this, we can add an arbitrary MIME Type by
writing the definition in the file of the path set in
the system property `content.types.user.table`.

Although it is possible to add arbitrary MIME types
even with mime-util which has been used so far, it was
judged that it is easier to understand according to the
method used in the standard library.