Removed dependency on mime-util
`MimeTypes` used mime-util to guess the MIME Type, but that
mime-util has not been maintained for a long time and now that
`java.net.URLConnection` can do the same thing.

Although mime-util uses magic.mime, the accuracy of guess is high,
but in generating content-type on the server side, there seems to
be no scenario where this is a problem ... as it becomes a problem,
we can set the content-type explicitly.

Also, mime-util uses extensions for MIME Type guesses, but can not
add any extensions.
`java.net.URLConnection` can add any extension to target.

For the above reason, replace with `java.net.URLConnection`