Excluded '#' and '&' from path separator
The parser generated by `SinatraPathPatternParser` generates
a regular expression `([^/?#]+)` for namedGroup.

With this regular expression, we can extract the character string up
to the front of `/` or `?` or `#`.

For example, the route `/file/to/scalatra.jar/hit` matches the route
definition `get("file/to/:path/hit")` and `params("path")` contains
`scalatra.jar`.

However, the source of the path is the `requestPath` method, and
`#` and `&` are removed in advance from the path returned by this
method (since the query and fragment are deleted from path).

In other words, `([^/]+)` is enough for the regular expression
to be generated.

The same applies to RailsPathPatternParser
(Rails has dot notation, so it becomes `([^/.]+)`)