Changed path encoding to be selectable
`requestPath` method was designed to decode Percent-Encoding except
for '/#?' used as a path delimiter.

After that, when acquiring it as a parameter, '/#?' are separately decoded.
In this case, however, there was a problem that a two-stage decoding resulted
in an unintended decoding result.

Also, it is very difficult to understand as a specification.

From the viewpoint of consistency, it is now possible to select whether
to decode or not.

There are two related issues:

https://github.com/scalatra/scalatra/issues/964
https://github.com/scalatra/scalatra/issues/969