Updated juniversalchardet
`juniversalchardet` has been updated to return US-ASCII more strictly and correctly.
However, the Content-Type charset set in the HTTP header should be UTF-8 as a standard.
Therefore, the change was made so that if juniversalchardet returns US-ASCII, it will be overwritten with UTF-8.

With this change, FileCharset will behave the same as before (UTF-8 will be preferred).