Changed URL generation method
In the submit method, the code always adds a slash between
the authority and the path in the URL generation.

Therefore, when a path beginning with a slash is passed to
the HTTP method, a path beginning with a double slash is
always generated.

This double slash was normalized in apache http client 4.x,
so it did not become apparent. However, in apache http client
5.x, the double slash is not normalized, so it may be judged
as an invalid URL in future updates of the library.

Therefore, the code was changed to determine whether the HTTP
method path begins with a slash or not and generate the correct
URL.

Although a valid path should start with a slash, for backward
compatibility, it is treated as a valid path even if it does
not start with a slash.